Refactor: greatly simplified types. Everything is now streamlined with the plotly Data type
Removed util functions to prepare the TimeSeriesData array into each
specific plot type data types. Removed the specific plot data types and
we now have a generic TimeSeriesData type with a child model/type named
PlotData that has the basic plotting properties (x, y and trace name
values).
We then include into each specific plot component a new function (will
be a hook) that handles the mapping of the TimeSeriesData objects into
an array of Data types that are plotly compliant.

We greatly reduced the number of types, simplified the code and leverage
plotlys types and intellisense with typescript